Module serde_keyvalue::key_values

Structs

Enum

KeyValueDeserializer

Serde deserializer for key-values strings.

ParseError

Error that may be thown while parsing a key-values string.

Enums

ErrorKind

Different kinds of errors that can be returned by the parser.

Functions

any_bool

Nom parser for booleans.

any_identifier

Nom parser for identifiers. An identifier may contain any alphanumeric character, as well as ‘_’ and ‘-’ at any place excepted the first one which cannot be ‘-’.

any_number

Nom parser for valid positive of negative numbers.

any_string

Nom parser for valid strings.

from_key_values

Attempts to deserialize T from the key-values string input.

Type Definitions

Result